feat: save offers fixed
